+++ ChangeLog.md
@@ -1,5 +1,13 @@
 Branch Android SDK change log
 
+- v2.5.0 Changing chrome tabs dependency optional. Adding option to simulate and test install. Few corner error  handling
+
+- v2.4.7 Fix for cold start issue with deeplink parameters on slower internet connection. Added local ip address as part of device params
+
+- v2.4.6 Fixed issue with receiving deep link prams through push notification while app is running in foreground. Added device locale info for stronger matching.
+
+- v2.4.5 Adding fix for transaction too large exception while auto deep linking. Fixing invalid argument exception while reading params from intent.
+
 - v2.4.4 Adding ability to skip collecting external intent data specific to URI host. Adding NPE protection for custom tab session access
 
 - v2.4.3 Adding ability to skip collecting external intent data specific to URI paths